requirements.txt
openai
datasets 
tqdm 
python-dotenv